+++ lib/screens/favorites_screen.dart
@@ -309,6 +309,11 @@ class _FavoritesScreenState extends ConsumerState<FavoritesScreen> {
   }
 
   Widget _buildFileThumbnail(VaultedFile file) {
+    if (!file.isEncrypted &&
+        (ref.watch(vaultSettingsProvider).value?.hideUnencryptedThumbnails ??
+            false)) {
+      return _buildPlaceholder(file);
+    }
     if (file.isImage) {
       if (file.isEncrypted) {
         return EncryptedThumbnail(file: file);